image {mixOmics}	R Documentation

Plot the cross-validation score.

Description

This function provide a image map (checkerboard plot) of the cross-validation score obtained by the tune.rcc function.

Usage

## S3 method for class 'tune.rcc'
plot(x, col = heat.colors, ...)

## S3 method for class 'tune.rcc'
image(x, col = heat.colors, ...)

Arguments

x
object returned by tune.rcc.

col
a character string specifying the colors function to use: terrain.colors, topo.colors, rainbow or similar functions. Defaults to heat.colors.

...
not used currently.

Details

plot.tune.rcc creates an image map of the matrix object$mat containing the cross-validation score obtained by the tune.rcc function. Also a color scales strip is plotted.

Value

none

Author(s)

Sébastien Déjean and Ignacio González.

See Also

tune.rcc, image.

Examples

data(nutrimouse)
X <- nutrimouse$lipid
Y <- nutrimouse$gene

## this can take some seconds
cv.score <- tune.rcc(X, Y, validation = "Mfold", plot = FALSE)
plot(cv.score)

# image(cv.score) # same result as plot()

[Package mixOmics version 6.10.8 Index]