imgCor {mixOmics}	R Documentation

Image Maps of Correlation Matrices between two Data Sets

Description

Display two-dimensional visualizations (image maps) of the correlation matrices within and between two data sets.

Usage

imgCor(
  X,
  Y,
  type = "combine",
  X.var.names = TRUE,
  Y.var.names = TRUE,
  sideColors = TRUE,
  interactive.dev = TRUE,
  title = TRUE,
  color,
  row.cex,
  col.cex,
  symkey,
  keysize,
  xlab,
  ylab,
  margins,
  lhei,
  lwid
)

Arguments

X
numeric matrix or data frame (n x p), the observations on the X variables. NAs are allowed.

Y
numeric matrix or data frame (n x q), the observations on the Y variables. NAs are allowed.

type
character string, (partially) maching one of "combine" or "separated", determining the kind of plots to be produced. See Details.

X.var.names, Y.var.names
logical, should the name of X- and/or Y-variables be shown? If TRUE (defaults) object$names$X and/or object$names$Y are used. Possible character vector with X- and/or Y-variable labels to use.

sideColors
character vector of length two. The color name for horizontal and vertical side bars that may be used to annotate the X and Y correlation matrices.

interactive.dev
Logical. The current graphics device that will be opened is interactive?

title
logical, should the main titles be shown?

color, xlab, ylab
arguments passed to cim.

row.cex, col.cex
positive numbers, used as cex.axis in for the row or column axis labeling. The defaults currently only use number of rows or columns, respectively.

symkey
Logical indicating whether the color key should be made symmetric about 0. Defaults to TRUE.

keysize
positive numeric value indicating the size of the color key.

margins
numeric vector of length two containing the margins (see par(mar)) for column and row names respectively.

lhei, lwid
arguments passed to layout to divide the device up into two rows and two columns, with the row-heights lhei and the column-widths lwid.

Details

If type="combine", the correlation matrix is computed of the combined matrices cbind(X, Y) and then plotted. If type="separate", three correlation matrices are computed, cor(X), cor(Y) and cor(X,Y) and plotted separately on a device. In both cases, a color correlation scales strip is plotted.

The correlation matrices are pre-processed before calling the image function in order to get, as in the numerical representation, the diagonal from upper-left corner to bottom-right one.

Missing values are handled by casewise deletion in the imgCor function.

If X.names = FALSE, the name of each X-variable is hidden. Default value is TRUE.

If Y.names = FALSE, the name of each Y-variable is hidden. Default value is TRUE.

Value

NULL (invisibly)

Author(s)

Ignacio González, Kim-Anh Lê Cao, Florian Rohart, Al J Abadi

See Also

cor, image, color.jet.

Examples

data(nutrimouse)
X <- nutrimouse$lipid
Y <- nutrimouse$gene

## 'combine' type plot (default)
imgCor(X, Y)

## Not run: 
## 'separate' type plot

imgCor(X, Y, type = "separate")

## 'separate' type plot without the name of datas
imgCor(X, Y, X.var.names = FALSE, Y.var.names = FALSE, type = "separate")

## End(Not run)

[Package mixOmics version 6.18.1 Index]