mint.pls {mixOmics}	R Documentation

P-integration

Description

Function to integrate and combine multiple independent studies measured on the same variables or predictors (P-integration) using variants of multi-group PLS (unsupervised analysis).

Usage

mint.pls(X,
Y,
ncomp = 2,
mode = c("regression", "canonical", "invariant", "classic"),
study,
scale = TRUE,
tol = 1e-06,
max.iter = 100,
near.zero.var = FALSE,
all.outputs = TRUE)

Arguments

X
numeric matrix of predictors combining multiple independent studies on the same set of predictors. NAs are allowed.

Y
Matrix or vector response for a multivariate regression framework. Data should be continuous variables (see mint.plsda for supervised classification and factor reponse)

ncomp
Number of components to include in the model (see Details). Default to 2

mode
character string. What type of algorithm to use, (partially) matching one of "regression" or "canonical". See Details.

study
factor indicating the membership of each sample to each of the studies being combined

scale
boleean. If scale = TRUE, each block is standardized to zero means and unit variances. Default = TRUE.

tol
Convergence stopping value.

max.iter
integer, the maximum number of iterations.

near.zero.var
boolean, see the internal nearZeroVar function (should be set to TRUE in particular for data with many zero values). Default = FALSE.

all.outputs
boolean. Computation can be faster when some specific (and non-essential) outputs are not calculated. Default = TRUE.

Details

mint.pls fits a vertical PLS-DA models with ncomp components in which several independent studies measured on the same variables are integrated. The aim is to explain the continuous outcome Y. The study factor indicates the membership of each sample in each study. We advise to only combine studies with more than 3 samples as the function performs internal scaling per study.

Multi (continuous)response are supported. X and Y can contain missing values. Missing values are handled by being disregarded during the cross product computations in the algorithm mint.pls without having to delete rows with missing data. Alternatively, missing data can be imputed prior using the nipals function.

The type of algorithm to use is specified with the mode argument. Four PLS algorithms are available: PLS regression ("regression"), PLS canonical analysis ("canonical"), redundancy analysis ("invariant") and the classical PLS algorithm ("classic") (see References and more details in ?pls).

Useful graphical outputs are available, e.g. plotIndiv, plotLoadings, plotVar.

Value

mint.pls returns an object of class "mint.pls", "pls", a list that contains the following components:

X
the centered and standardized original predictor matrix.

Y
the centered and standardized original response vector or matrix.

ncomp
the number of components included in the model.

study
The study grouping factor

mode
the algorithm used to fit the model.

variates
list containing the variates of X - global variates.

loadings
list containing the estimated loadings for the variates - global loadings.

variates.partial
list containing the variates of X relative to each study - partial variates.

loadings.partial
list containing the estimated loadings for the partial variates - partial loadings.

names
list containing the names to be used for individuals and variables.

nzv
list containing the zero- or near-zero predictors information.

iter
Number of iterations of the algorthm for each component

explained_variance
Percentage of explained variance for each component and each study (note that contrary to PCA, this amount may not decrease as the aim of the method is not to maximise the variance, but the covariance between data sets).

Author(s)

Florian Rohart, Kim-Anh Lê Cao

References

Rohart F, Eslami A, Matigian, N, Bougeard S, Lê Cao K-A (2017). MINT: A multivariate integrative approach to identify a reproducible biomarker signature across multiple experiments and platforms. BMC Bioinformatics 18:128.

Eslami, A., Qannari, E. M., Kohler, A., and Bougeard, S. (2014). Algorithms for multi-group PLS. J. Chemometrics, 28(3), 192-201.

See Also

spls, summary, plotIndiv, plotVar, predict, perf, mint.spls, mint.plsda, mint.splsda and http://www.mixOmics.org/mixMINT for more details.

Examples

data(stemcells)

# for the purpose of this example, we artificially
# create a continuous response Y by taking gene 1.

res = mint.pls(X = stemcells$gene[,-1], Y = stemcells$gene[,1], ncomp = 3,
study = stemcells$study)

plotIndiv(res)

#plot study-specific outputs for all studies
plotIndiv(res, study = "all.partial")

## Not run: 
#plot study-specific outputs for study "2"
plotIndiv(res, study = "2", col = 1:3, legend = TRUE)

## End(Not run)

[Package mixOmics version 6.6.2 Index]