plot.perf.pls {mixOmics}	R Documentation

Plot for model performance for PLS analyses

Description

Function to plot performance criteria, such as MSEP, RMSEP, R^2, Q^2 for s/PLS methods as a function of the number of components.

Usage

## S3 method for class 'perf.pls.mthd'
plot(
  x,
  criterion = "MSEP",
  xlab = "Number of components",
  ylab = NULL,
  LimQ2 = 0.0975,
  LimQ2.col = "grey30",
  sd = NULL,
  pch = 1,
  pch.size = 3,
  cex = 1.2,
  col = color.mixo(1),
  title = NULL,
  ...
)

## S3 method for class 'perf.spls.mthd'
plot(
  x,
  criterion = "MSEP",
  xlab = "Number of components",
  ylab = NULL,
  LimQ2 = 0.0975,
  LimQ2.col = "grey30",
  sd = NULL,
  pch = 1,
  pch.size = 3,
  cex = 1.2,
  col = color.mixo(1),
  title = NULL,
  ...
)

Arguments

x
an perf.pls object.

criterion
character string. What type of validation criterion to plot for pls or spls. One of "MSEP", "RMSEP", "R2" or "Q2". More measures available for pls2 methods. See perf.

xlab, ylab
titles for x and y axes. Typically character strings, but can be expressions (e.g., expression(R^2)).

LimQ2
numeric value. Signification limit for the components in the model. Default is LimQ2 = 0.0975.

LimQ2.col
character string specifying the color for the LimQ2 line to be plotted. If "none" the line will not be plotted.

sd
If 'nrepeat' was used in the call to 'perf', error bar shows the standard deviation if sd=TRUE. For mint objects sd is set to FALSE as the number of repeats is 1.

pch
Plot character to use.

pch.size
Plot character size to use.

cex
A numeric which adjusts the font size in the plot.

col
Character. Colour to be used for data points.

title
Character, Plot title. Not used by PLS2 feature-wise measure plots.

...
Not used.

Details

plot.perf creates one plot for each response variable in the model, laid out in a multi-panel display. See ?perf for examples.

Value

none

Author(s)

Al J Abadi

References

Rohart F, Gautier B, Singh A, Lê Cao K-A. mixOmics: an R package for 'omics feature selection and multiple data integration. PLoS Comput Biol 13(11): e1005752

See Also

pls, spls, plsda, splsda, perf.

[Package mixOmics version 6.18.1 Index]