plot.rcc {mixOmics}	R Documentation

Canonical Correlations Plot

Description

This function provides scree plot of the canonical correlations.

Usage

## S3 method for class 'rcc'
plot(x, type = "barplot", ...)

Arguments

x
object of class inheriting from "rcc".

type
Character, default "barplot" or any other type available in plot, as "l","b","p",..

...
Not used

Value

none

Author(s)

Sébastien Déjean, Ignacio González, Al J Abadi

See Also

points, barplot, par.

Examples

data(nutrimouse)
X <- nutrimouse$lipid
Y <- nutrimouse$gene
nutri.res <- rcc(X, Y, lambda1 = 0.064, lambda2 = 0.008)

## 'pointplot' type scree
plot(nutri.res) #(default)

## Not run: 
plot(nutri.res, pch = 19, cex = 1.2,
col = c(rep("red", 3), rep("darkblue", 18)))

## 'barplot' type scree
plot(nutri.res, type = "barplot")

plot(nutri.res, type = "barplot", density = 20, col = "black")

## End(Not run)

[Package mixOmics version 6.18.1 Index]