plotMarkers {mixOmics}	R Documentation

Plot the values for multivariate markers in block analyses

Description

Plots the standardised values (after centring and/or scaling) for the selected variables for a given block on a given component. Only applies to block.splsda or block.spls.

Usage

plotMarkers(
  object,
  block,
  markers = NULL,
  comp = 1,
  group = NULL,
  col.per.group = NULL,
  global = FALSE,
  title = NULL,
  violin = TRUE,
  boxplot.width = NULL,
  violin.width = 0.9
)

Arguments

object
An object of class block.splsda or block.spls

block
Name or index of the block to use

markers
Character or integer, only include these markers. If integer, the top 'markers' features are shown

comp
Integer, the component to use

group
Factor, the grouping variable (only required for block.spls objects)

col.per.group
character (or symbol) color to be used when 'group' is defined. Vector of the same length as the number of groups.

global
Logical indicating whether to show the global plots (TRUE) or segregate by feature (FALSE). Only available when object$scale=TRUE

title
The plot title

violin
(if global = FALSE) Logical indicating whether violin plots should also be shown

boxplot.width
Numeric, adjusts the width of the box plots

violin.width
Numeric, adjusts the width of the violin plots

Value

A ggplot object

See Also

plotLoadings, block.splsda, block.spls

Examples

# see ?block.splsda and ?block.spls

[Package mixOmics version 6.18.1 Index]