summary {mixOmics}	R Documentation

Summary Methods for CCA and PLS objects

Description

Produce summary methods for class "rcc", "pls" and "spls".

Usage

## S3 method for class 'rcc'
summary(object, what = c("all", "communalities", "redundancy"),
cutoff = NULL, digits = 4, ...)

## S3 method for class 'mixo_pls'
summary(object, what = c("all", "communalities", "redundancy",
"VIP"), digits = 4, keep.var = FALSE, ...)

## S3 method for class 'mixo_spls'
summary(object, what = c("all", "communalities", "redundancy",
"VIP"), digits = 4, keep.var = FALSE, ...)

Arguments

object
object of class inherited from "rcc", "pls" or "spls".

cutoff
real between 0 and 1. Variables with all correlations components below this cutoff in absolute value are not showed (see Details).

digits
integer, the number of significant digits to use when printing. Defaults to 4.

what
character string or vector. Should be a subset of c("all", "summarised", "communalities", "redundancy", "VIP"). "VIP" is only available for (s)PLS. See Details.

keep.var
boolean. If TRUE only the variables with loadings not zero (as selected by spls) are showed. Defaults to FALSE.

...
not used currently.

Details

The information in the rcc, pls or spls object is summarised, it includes: the dimensions of X and Y data, the number of variates considered, the canonical correlations (if object of class "rcc") and the (s)PLS algorithm used (if object of class "pls" or "spls") and the number of variables selected on each of the sPLS components (if x of class "spls").

"communalities" in what gives Communalities Analysis. "redundancy" display Redundancy Analysis. "VIP" gives the Variable Importance in the Projection (VIP) coefficients fit by pls or spls. If what is "all", all are given.

For class "rcc", when a value to cutoff is specified, the correlations between each variable and the equiangular vector between X- and Y-variates are computed. Variables with at least one correlation componente bigger than cutoff are showed. The defaults is cutoff=NULL all the variables are given.

Value

The function summary returns a list with components:

ncomp
the number of components in the model.

cor
the canonical correlations.

cutoff
the cutoff used.

keep.var
list containing the name of the variables selected.

mode
the algoritm used in pls or spls.

Cm
list containing the communalities.

Rd
list containing the redundancy.

VIP
matrix of VIP coefficients.

what
subset of c("all", "communalities", "redundancy", "VIP").

digits
the number of significant digits to use when printing.

method
method used: rcc, pls or spls.

Author(s)

Sébastien Déjean, Ignacio González and Kim-Anh Lê Cao.

See Also

rcc, pls, spls, vip.

Examples

## summary for objects of class 'rcc'
data(nutrimouse)
X <- nutrimouse$lipid
Y <- nutrimouse$gene
nutri.res <- rcc(X, Y, ncomp = 3, lambda1 = 0.064, lambda2 = 0.008)
more <- summary(nutri.res, cutoff = 0.65)

## Not run: 
## summary for objects of class 'pls'
data(linnerud)
X <- linnerud$exercise
Y <- linnerud$physiological
linn.pls <- pls(X, Y)
more <- summary(linn.pls)

## summary for objects of class 'spls'
data(liver.toxicity)
X <- liver.toxicity$gene
Y <- liver.toxicity$clinic
toxicity.spls <- spls(X, Y, ncomp = 3, keepX = c(50, 50, 50),
keepY = c(10, 10, 10))
more <- summary(toxicity.spls, what = "redundancy", keep.var = TRUE)

## End(Not run)

[Package mixOmics version 6.6.2 Index]