yeast {mixOmics}	R Documentation

Yeast metabolomic study

Description

Two Saccharomyces Cerevisiae strains were compared under two different environmental conditions, 37 metabolites expression are measured.

Usage

data(yeast)

Format

A list containing the following components:

data

data matrix with 55 rows and 37 columns. Each row represents an experimental sample, and each column a single metabolite.

strain

a factor containing the type of strain (MT or WT).

condition

a factor containing the type of environmental condition (AER or ANA).

strain.condition

a crossed factor between strain and condition.

Details

In this study, two Saccharomyces cerevisiae strains were used - wild-type (WT) and mutant (MT), and were carried out in batch cultures under two different environmental conditions, aerobic (AER) and anaerobic (ANA) in standard mineral media with glucose as the sole carbon source. After normalization and pre processing, the metabolomic data results in 37 metabolites and 55 samples which include 13 MT-AER, 14 MT-ANA, 15 WT-AER and 13 WT-ANA samples

Value

none

References

Villas-Boas S, Moxley J, Akesson M, Stephanopoulos G, Nielsen J: High-throughput metabolic state analysis (2005). The missing link in integrated functional genomics. Biochemical Journal, 388:669–677.

[Package mixOmics version 6.10.8 Index]