explained_variance {mixOmics}	R Documentation

Calculation of explained variance

Description

This function calculates the variance explained by variates.

Usage

explained_variance(data, variates, ncomp)

Arguments

data
numeric matrix of predictors

variates
variates as obtained from a pls object for instance

ncomp
number of components. Should be lower than the number of columns of variates

Details

explained_variance calculates the explained variance of each variates out of the total variance in data.

Value

explained_variance simply returns the explained variance for each variate.

Author(s)

Florian Rohart

See Also

spls, splsda, plotIndiv, plotVar, cim, network.

Examples

data(liver.toxicity)
X <- liver.toxicity$gene
Y <- liver.toxicity$clinic

toxicity.spls <- spls(X, Y, ncomp = 2, keepX = c(50, 50), keepY = c(10, 10))

ex = explained_variance(toxicity.spls$X, toxicity.spls$variates$X, ncomp =2)

# ex should be the same as
toxicity.spls$explained_variance$X

[Package mixOmics version 6.6.2 Index]