plot.rcc {mixOmics}	R Documentation

Canonical Correlations Plot

Description

This function provides scree plot of the canonical correlations.

Usage

## S3 method for class 'rcc'
plot(x, scree.type = c("pointplot", "barplot"), ...)

Arguments

x
object of class inheriting from "rcc".

scree.type
character string, (partially) matching one of "pointplot" or "barplot", determining the kind of scree plots to be produced.

...
arguments to be passed to other methods. For the "pointplot" type see points, for "barplot" type see barplot.

Value

none

Author(s)

Sébastien Déjean and Ignacio González.

See Also

points, barplot, par.

Examples

data(nutrimouse)
X <- nutrimouse$lipid
Y <- nutrimouse$gene
nutri.res <- rcc(X, Y, lambda1 = 0.064, lambda2 = 0.008)

## 'pointplot' type scree
plot(nutri.res) #(default)

## Not run: 
plot(nutri.res, pch = 19, cex = 1.2,
col = c(rep("red", 3), rep("darkblue", 18)))

## 'barplot' type scree
plot(nutri.res, scree.type = "barplot")

plot(nutri.res, scree.type = "barplot", density = 20, col = "black")

## End(Not run)

[Package mixOmics version 6.6.2 Index]