srbct {mixOmics}	R Documentation

Small version of the small round blue cell tumors of childhood data

Description

This data set from Khan et al., (2001) gives the expression measure of 2308 genes measured on 63 samples.

Usage

data(srbct)

Format

A list containing the following components:

gene

data frame with 63 rows and 2308 columns. The expression measure of 2308 genes for the 63 subjects.

class

A class vector containing the class tumour of each case (4 classes in total).

gene.name

data frame with 2308 rows and 2 columns containing further information on the genes.

Value

none

Source

http://research.nhgri.nih.gov/microarray/Supplement

References

Khan et al. (2001). Classification and diagnostic prediction of cancers using gene expression profiling and artificial neural networks. Nature Medicine 7, Number 6, June.

[Package mixOmics version 6.10.8 Index]